+++ projects/ccd-case-ui-toolkit/src/lib/shared/components/palette/case-flag/enums/write-case-flag-field.enum.ts
@@ -14,3 +14,8 @@ export enum CaseFlagText {
   CAPTION_EXTERNAL = 'Request support',
   CAPTION_NONE = ''
 }
+
+export enum CaseFlagErrorMessage {
+  NO_EXTERNAL_FLAGS_COLLECTION = 'External collection for storing this case flag has not been configured for this case type',
+  NO_INTERNAL_FLAGS_COLLECTION = 'Internal collection for storing this case flag has not been configured for this case type'
+}